requirements.txt
PyGObject>=3.46.0
GitPython>=3.1.37
